custom alert

+++ lib/ui/screens/home/sidebar/settings.dart
@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
+import 'package:podlove_flutter/constants/app_colors.dart';
 import 'package:podlove_flutter/providers/auth/delete_account_provider.dart';
 import 'package:podlove_flutter/routes/route_path.dart';
 import 'package:podlove_flutter/ui/widgets/custom_app_bar.dart';
 import 'package:podlove_flutter/ui/widgets/custom_outline_button.dart';
+import 'package:podlove_flutter/ui/widgets/custom_text.dart';
 
 class Settings extends ConsumerWidget {
   const Settings({super.key});
@@ -15,7 +17,7 @@ class Settings extends ConsumerWidget {
     final deleteAccountNotifier = ref.read(deleteAccountProvider.notifier);
 
     ref.listen<DeleteAccountState>(deleteAccountProvider, (prev, current) {
-      if (current.isSuccess == true) {
+      if (current.isSuccess == true && current.isLoading == false) {
         context.go(RouterPath.signIn);
       }
     });
@@ -48,8 +50,52 @@ class Settings extends ConsumerWidget {
                       : 'Delete Account',
                   onPressed: deleteAccountState.isLoading
                       ? null
-                      : () async {
-                          deleteAccountNotifier.deleteAccount();
+                      : () {
+                          showDialog(
+                            context: context,
+                            builder: (BuildContext context) {
+                              return AlertDialog(
+                                title: CustomText(
+                                  text: "Are you sure?",
+                                  color: AppColors.primaryText,
+                                ),
+                                content: const Text(
+                                  'Are you sure you want to delete your account?',
+                                ),
+                                actions: <Widget>[
+                                  TextButton(
+                                    style: TextButton.styleFrom(
+                                      textStyle: Theme.of(context)
+                                          .textTheme
+                                          .labelLarge,
+                                    ),
+                                    child: CustomText(
+                                      text: "Nevermind",
+                                      color: AppColors.primaryText,
+                                    ),
+                                    onPressed: () {
+                                      Navigator.pop(context);
+                                    },
+                                  ),
+                                  TextButton(
+                                    style: TextButton.styleFrom(
+                                      textStyle: Theme.of(context)
+                                          .textTheme
+                                          .labelLarge,
+                                    ),
+                                    child: CustomText(
+                                      text: "Yes",
+                                      color: AppColors.primaryText,
+                                    ),
+                                    onPressed: () async {
+                                      deleteAccountNotifier.deleteAccount();
+                                      Navigator.pop(context);
+                                    },
+                                  ),
+                                ],
+                              );
+                            },
+                          );
                         },
                 ),
               ],
